minor changes to make everything compile cleanly with "make strict".  Also fiddle with Makefile.am so that init.source.c isn't compiled.

+++ monte/gsl_monte_vegas.h
@@ -25,7 +25,7 @@ typedef struct {
   int bins_prev;
   int calls_per_box;
   int it_num;
-  int num_dim;
+  unsigned long num_dim;
   int bins;
   int boxes;
   int init_done;
